requirements.txt
google-genai>=1.0.0
google-cloud-storage>=2.16.0
fastapi>=0.111.0
uvicorn>=0.30.0
selenium>=4.21.0
Pillow>=10.0.0
python-multipart>=0.0.9
httpx>=0.28.1
python-dotenv>=1.0.0
jinja2>=3.1.0
pydantic>=2.7.0
slowapi>=0.1.9
pytest>=8.0.0
pytest-asyncio>=0.23.0

